Move theme provider into src dir
