+++ theme-provider/index.ts
@@ -1,1 +1,3 @@
-export * from './context/ThemeProvider';
+export { useTheme, ThemeProvider } from './src/context/ThemeProvider';
+export { ThemeContextConfig } from './src/types/themeContextConfig';
+export { WhiteLabelColors } from './src/types/whiteLabelColors';
