+++ theme-provider/src/utils/composeTheme.ts
@@ -1,7 +1,7 @@
 import { WhiteLabelColors } from '../types/whiteLabelColors';
-import { ThemeBase } from '../../theme';
+import { ThemeBase } from '../../../theme';
 import { ConfigType } from '@stitches/react/types/config';
-import { themeList } from '../../theme-list';
+import { themeList } from '../../../theme-list';
 
 export const composeTheme = (themeBase: ThemeBase, colors: WhiteLabelColors = {}): ConfigType.Theme => ({
   ...themeList[themeBase],
